Add remove documentation and parameter

+++ remove.sh
@@ -6,4 +6,7 @@ BASEDIR="$(dirname ${REALPATH})"
 
 docker image rm "${DOCKER_IMAGE_NAME}"
 docker volume rm ${DOCKER_ATOM_HOME_VOLUME_NAME}
-docker volume rm ${DOCKER_RUST_HOME_VOLUME_NAME}
+
+if [ "$1" != "--keep-rust-home-volume" ]; then
+  docker volume rm ${DOCKER_RUST_HOME_VOLUME_NAME}
+fi
